+++ Example/Pods/Headers/Private/KFXAdditions/FoundationAdditions.h
@@ -1,0 +1,1 @@
+../../../../../KFXAdditions/Classes/Foundation/FoundationAdditions.h