+++ KFXAdditions/Classes/KFXAdditions.h
@@ -1,0 +1,19 @@
+/********************************
+ *
+ * Copyright © 2016-2017 Christian Fox
+ * All Rights Reserved
+ * Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ *
+ * This file is included with KFXAdditions
+ *
+ ************************************/
+
+#ifndef KFXAdditions_h
+#define KFXAdditions_h
+
+#import <KFXAdditions/FoundationAdditions.h>
+#import <KFXAdditions/UIKitAdditions.h>
+#import <KFXAdditions/CLLocation+KFXAdditions.h>
+
+
+#endif /* KFXAdditions_h */